requirements.txt
google-cloud-storage
langchain==0.0.346
pyhcl
google-generativeai
google-cloud-aiplatform
python-dotenv
google-generativeai
google-search-results
duckduckgo-search
google-auth
pypdf
google-cloud-storage
unstructured
google-cloud-documentai
google-cloud-documentai-toolbox
streamlit==1.27.2
streamlit-chat
lark
tiktoken
pinecone-client
cohere
tfds-nightly
tensorflow_hub==0.13.0
jq